# openapi/integrator-api.yaml
openapi: 3.1.0
info:
  title: Vlantegrator API
  description: API платформы Vlantegrator
  contact:
    name: Support
    email: vlantegrator-support@gmail.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/v1/user/settings:
    get:
      tags:
      - Routes API
      summary: Получить настройки пользователя
      description: Получить настройки пользователя
      operationId: getUserSettings
      responses:
        "200":
          description: Настройки пользователя возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettingsDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    post:
      tags:
      - Routes API
      summary: Обновить настройки пользователя
      description: Обновить настройки пользователя
      operationId: updateUserSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsDto"
        required: true
      responses:
        "200":
          description: Настройки пользователя обновлены
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/route:
    get:
      tags:
      - Routes API
      summary: Получить информацию о всех маршрутах
      description: Отдаёт информацию о всех маршрутах
      operationId: getAllRoutes
      responses:
        "200":
          description: Информация о всех маршрутах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    post:
      tags:
      - Routes API
      summary: Создать маршрут
      description: Создаёт новый маршрут на основе переданной конфигурации
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequestDto"
        required: true
      responses:
        "201":
          description: Маршрут создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteIdDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/route/{id}/{versionHash}/definition:
    get:
      tags:
      - Routes API
      summary: Получить структуру маршрута
      description: Отдаёт структуру маршрута в качестве ответа
      operationId: getRouteDefinition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      - name: versionHash
        in: path
        required: true
        schema:
          type: string
          description: Хэш-код версии маршрута
      responses:
        "200":
          description: Структура маршрута получена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDefinitionDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    post:
      tags:
      - Routes API
      summary: Обновить структуру маршрута
      description: Обновляет структуру имеющегося маршрута на основе переданной конфигурации
      operationId: updateRouteDefinition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      - name: versionHash
        in: path
        required: true
        schema:
          type: string
          description: Хэш-код версии маршрута
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteDefinitionDto"
        required: true
      responses:
        "200":
          description: Структура маршрута обновлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteIdDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/route/{id}/deploy:
    post:
      tags:
      - Routes API
      summary: Развернуть маршрут
      description: "Выполнить действие по отношению к развёртыванию маршрута: запу\
        стить, остановить, удалить, перезапустить"
      operationId: deployRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      - name: action
        in: query
        required: true
        schema:
          type: string
          description: Действие
      responses:
        "202":
          description: Запрос на выполнение действия по отношению к развёртыванию
            маршрута принят
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/route/{id}/build:
    post:
      tags:
      - Routes API
      summary: Асинхронная сборка маршрута
      description: Запускает процесс сборки маршрута (compile + build image) в фоновом
        режиме
      operationId: buildRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      responses:
        "202":
          description: Запрос на сборку принят
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  /api/v1/route/search:
    post:
      tags:
      - Routes API
      summary: Поиск информации о маршрутах по критерию
      description: "Отдаёт информацию о маршрутах, соответствующих критерию"
      operationId: searchRoutes
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRoutesRequestDto"
        required: true
      responses:
        "200":
          description: "Информация о маршрутах, соответствующих критерию"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/route/{id}:
    get:
      tags:
      - Routes API
      summary: Получить информацию о маршруте
      description: Отдаёт информацию о маршруте
      operationId: getRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      responses:
        "200":
          description: Информация о маршруте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Routes API
      summary: Удалить маршрут
      description: "Полностью удалить маршрут, в том числе все его контейнер и обр\
        азы"
      operationId: deleteRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      responses:
        "204":
          description: Маршрут удалён
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Routes API
      summary: Обновить конфиг маршрута
      description: Обновляет конфиг имеющегося маршрута на основе переданной конфигурации
      operationId: updateRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: ID маршрута
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouteRequestDto"
        required: true
      responses:
        "200":
          description: Конфиг маршрута обновлён
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/route/actions:
    get:
      tags:
      - Routes API
      summary: Получить действия пользователей над маршрутами
      description: "Получить список действий, предпринимаемых текущим или всеми по\
        льзователями над маршрутами в платформе, отсортированный по времени соверш\
        ения действия от нового к старому"
      operationId: getRouteUserActions
      parameters:
      - name: personal
        in: query
        required: true
        schema:
          type: boolean
          description: Возвращать только действия текущего пользователя
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          description: Ограничить количество возвращаемых записей
      responses:
        "200":
          description: Список действий над маршрутами
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteUserActionDto"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    UserAccessibilitySettingsDto:
      type: object
      description: Пользовательские настройки accessibility
      properties:
        disableAnimations:
          type: boolean
          description: Выключить анимации
        enableHighContrast:
          type: boolean
          description: Включить режим высокой контрастности
      required:
      - disableAnimations
      - enableHighContrast
...
    UserSettingsDto:
      type: object
      description: Пользовательские настройки. Не влияют на работу всей платформы
      properties:
        editor:
          $ref: "#/components/schemas/UserEditorSettingsDto"
          description: Настройки редактора маршрутов
        accessibility:
          $ref: "#/components/schemas/UserAccessibilitySettingsDto"
          description: Настройки редактора accessibility
      required:
      - accessibility
      - editor
    RouteIdDto:
      type: object
      description: Полный идентификатор маршрута
      properties:
        id:
          type: string
          format: uuid
          description: ID маршрута
        versionHash:
          type: string
          description: Хэш-код версии маршрута
      required:
      - id
    CreateRouteRequestDto:
      type: object
      description: Запрос на создание маршрута
      properties:
        name:
          type: string
          description: Имя маршрута
        description:
          type: string
          description: Описание маршрута
        ownerName:
          type: string
          description: Username владельца. Если пустой – владельцем маршрута становится
            инициатор запроса
        publishedPorts:
          type: string
          description: Маппинг портов
        networks:
          type: array
          description: Сети маршрута. Должны быть предварительно созданы
          items:
            $ref: "#/components/schemas/DockerNetworkDto"
        env:
          type: object
          additionalProperties:
            type: object
          description: Конфигурация среды маршрута
      required:
      - env
      - name
      - networks
...
    RouteDefinitionDto:
      type: object
      description: Структура маршрута
      properties:
        nodes:
          type: array
          description: Узлы маршрута
          items:
            $ref: "#/components/schemas/NodeDto"
        connections:
          type: array
          description: Соединения маршрута
          items:
            $ref: "#/components/schemas/ConnectionDto"
      required:
      - connections
      - nodes
    RouteDto:
      type: object
      description: Конфигурация маршрута
      properties:
        routeId:
          $ref: "#/components/schemas/RouteIdDto"
          description: Идентификатор маршрута (uuid + versionHash)
        name:
          type: string
          description: Имя маршрута
        description:
          type: string
          description: Описание маршрута
        ownerName:
          type: string
          description: Username владельца. Если пустой – владельцем маршрута становится
            инициатор запроса
        publishedPorts:
          type: string
          description: Маппинг портов
        networks:
          type: array
          description: Сети маршрута. Должны быть предварительно созданы
          items:
            $ref: "#/components/schemas/DockerNetworkDto"
        env:
          type: object
          additionalProperties:
            type: object
          description: Конфигурация среды маршрута
      required:
      - description
      - env
      - name
      - networks
      - ownerName
      - publishedPorts
      - routeId
    SearchRoutesRequestDto:
      type: object
      description: Запрос на поиск маршрутов
      properties:
        field:
          type: string
          description: "Поле, по которому производится поиск"
          enum:
          - ID
          - NAME
          - OWNER
        query:
          type: string
          description: "Значение, введённое пользователем в строке поиска"
          minLength: 1
      required:
      - field
      - query
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT